Enhance project structure and documentation
- Updated .gitignore to include additional build artifacts and configuration files.
- Modified CMakeLists.txt to include a new validation test source file and adjust include directories.
- Expanded README.md with detailed build, run, and validation instructions, including a new section for performance benchmarks.
- Added new validation and known issues documentation to improve user guidance.
- Introduced new JSON input files for validation tests and updated data structures to support double precision for certain parameters.
- Enhanced Django views to dynamically locate the C++ executable and handle simulation requests more robustly.
- Added scripts for building and plotting validation results, improving usability for end-users.
- Updated CI configuration to ensure tests run on both Linux and Windows environments.

+++ lib/include/DataStructs.h
@@ -20,8 +20,8 @@ namespace DATA_STRUCTS {
 		double thermalVelocity;
 		double spatialPerturbationAmplitude;
 		int spatialPerturbationMode;
-		int plasmaFrequency;
-		int chargeMassRatio;
+		double plasmaFrequency;
+		double chargeMassRatio;
 		double particleCharge;
 		double particleMass;
 		double chargeCloudWidth;
